refactor(i18n): move app tree under [locale] and add locale layout
- git mv page.tsx, not-found.tsx, story/, submit/, admin/ under [locale]/
- Replace root layout.tsx with minimal pass-through (no <html>)
- Create [locale]/layout.tsx with NextIntlClientProvider, setRequestLocale,
  generateStaticParams, hasLocale guard, and fonts (Bodoni_Moda latin+latin-ext,
  Space_Grotesk latin)
- Fix action imports in admin/page.tsx, admin/story/[id]/page.tsx,
  AdminLoginForm.tsx, SubmitForm.tsx to use new [locale] paths
