+++ src/app/[locale]/story/[slug]/page.tsx
