+++ src/app/[locale]/submit/page.tsx
