+++ src/app/[locale]/submit/thanks/page.tsx
