+++ src/app/layout.tsx
@@ -1,31 +1,9 @@
 import "./globals.css";
 import type { ReactNode } from "react";
-import { Bodoni_Moda, Space_Grotesk } from "next/font/google";
 
-const serif = Bodoni_Moda({
-  subsets: ["latin"],
-  weight: ["400", "500", "600"],
-  style: ["normal", "italic"],
-  variable: "--font-serif",
-  display: "swap",
-});
-
-const sans = Space_Grotesk({
-  subsets: ["latin"],
-  weight: ["400", "500", "600"],
-  variable: "--font-sans",
-  display: "swap",
-});
-
-export const metadata = {
-  title: "Nocturne — Архив интернет-хоррора",
-  description: "Архив интернет-хоррора: истории, сущности и городские легенды сети.",
-};
-
+// The real <html>/<body> live in app/[locale]/layout.tsx so the lang attribute
+// can reflect the active locale. This root layout only satisfies Next's
+// requirement for a top-level layout and imports global styles once.
 export default function RootLayout({ children }: { children: ReactNode }) {
-  return (
-    <html lang="ru" className={`${serif.variable} ${sans.variable}`}>
-      <body className="bg-bg font-sans text-ink antialiased">{children}</body>
-    </html>
-  );
+  return children;
 }
